requirements.txt
hydra-core
torch
whisper
pyannote.audio
scikit-learn
loguru